+++ changelog.md
@@ -1,3 +1,7 @@
+1.0.4:
+* Changed: Recipe for Amelioration Kit again
+* Added: Tier IV Bonus Modifier
+
 1.0.3:
 * Changed: Auto-Repair is now Amelioration (a synonym for "improving"). The mechanics work the same, but recipe has changed
 * Changed: Textures for all items + proper 1.9.x+ model support
@@ -11,10 +15,13 @@
 
 Where [M] is Moss Stone, and E is an enchanted book or item with Mending on it
 
+
 1.0.2:
 * Fixed: Auto-Repair interrupting block break / item use
+
 1.0.1:
 * Fixed: Auto-Repair Working on Multiple Tools Simultaneously
+
 1.0:
 Initial Release:
 * Added: Auto-Repair Modifier